+++ codegen/src/generate/mod.rs
@@ -57,5 +57,9 @@ pub(crate) fn generate_impl(mode: &Mode, params: Parameters) -> Result<TokenStre
         Mode::Merge => Ok(merge::owned::impl_owned(&params, similar_fields)),
         Mode::MergeRef => Ok(merge::borrowed::impl_borrowed(&params, similar_fields)),
         Mode::Into => Ok(into::normal::impl_into(&params, similar_fields)),
+        Mode::IntoDefault => Ok(into::with_default::impl_into_default(
+            &params,
+            similar_fields,
+        )),
     }
 }