+++ codegen/src/lib.rs
@@ -19,6 +19,7 @@ enum Mode {
     Merge,
     MergeRef,
     Into,
+    IntoDefault,
 }
 
 pub(crate) struct Parameters {
@@ -55,7 +56,7 @@ pub(crate) struct Parameters {
 ///
 /// Eiter a single path or a list of paths can be specified.
 /// The traits will then be implemented for each given target struct.
-#[proc_macro_derive(InterStruct, attributes(merge, merge_ref, into))]
+#[proc_macro_derive(InterStruct, attributes(merge, merge_ref, into, into_default))]
 pub fn inter_struct(struct_ast: TokenStream) -> TokenStream {
     // Parse the main macro input as a struct.
     // We work on a clone of the struct ast.
@@ -84,6 +85,7 @@ pub fn inter_struct(struct_ast: TokenStream) -> TokenStream {
             "merge" => Mode::Merge,
             "merge_ref" => Mode::MergeRef,
             "into" => Mode::Into,
+            "into_default" => Mode::IntoDefault,
             _ => continue,
         };